+++ src/main/java/com/intranet/repository/TimeSheetRepo.java
@@ -8,5 +8,5 @@ import com.intranet.entity.TimeSheet;
 
 public interface TimeSheetRepo extends JpaRepository<TimeSheet, Long> {
 
-    List<TimeSheet> findByUserIdOrderByWorkDateDesc(Integer userId);
+    List<TimeSheet> findByUserIdOrderByWorkDateDesc(Long userId);
 }
